UD06_Redis/CarTracker/requirements.txt
redis
python-dotenv
pymongo[srv]
psycopg2
websockets
tkintermapview